backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
openai>=1.0.0
pgvector>=0.2.0
sentence-transformers>=2.2.0
datasets>=2.14.0
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
